Added m1 and 1p like functions for numerical stability
Added m1 and 1p like functions to discount, compound and implied rate methods. The minus one and one plus function help avoid loss of accuracy in certain contexts, and is akin to the expm1 and log1p in C++ std and numpy. It is used to improve accuracte for the equivalentRate function, but can have other uses.